JSEXT1/C/compile.js: inline JSEXT1.trim(), and simplify
The code is looking for compile errors, by peeking at the output of gcc.  For some reason it used to split it into lines, grep each line for "error:", and then throw the entire output (trimmed) rather than just that line.  Since it doesn't appear to matter much, the new version just greps the entire thing for "error:", and then throws a proper Error object