replace JSEXT1.c.cpp() with a new .C.runcpp(), that forks GCC's cpp, rather than using our custom libcpp
The .jswrapper files created with the new code are slightly different (they have a few more #define'd constants included), but not in any ways that seem to matter.

runcpp() uses JSEXT1.File's popen() stuff, so 0-makeclib.js needs more hard-coded wrappers for libc stuff, to make that work

runcpp() doesn't have the optional argument for additional include paths, because no-one was using it.