make JSX_ReportException() just a macro wrapping JS_ReportError()
Previously this didn't work, as our error reporter would output the error to stdout even though it was being handled as an exception.  This no longer happens, possibly due to upgrading SpiderMonkey from 1.6 to 1.8b1.  (And even if it did, adding a & JSREPORT_EXCEPTION would prevent it).

We can't make JSX_ReportException() a wrapper *function*, because you call a varargs function like JS_ReportError() using a va_list.